Adds rbac configuration for purposes of using in combination with community nginx ingress controller
The community helm chart for nginx ingress is not configured with the correct RBAC permissions to be used in a cluster using RBAC. It is preferable to use the community chart, as app registry is being deprecated. While waiting to submit and navigate an upsteam pull request to the community chart, we are packaging the RBAC permissions with this app so that it can be operated and run in any cluster from anywhere, in combination with the nginx ingress controller stable/nginx-ingress. If and when the community chart supports RBAC permissions, this template will be removed.